fix loyalty points errors

+++ nasnav-services/src/main/java/com/nasnav/service/CoinsDropService.java
@@ -13,7 +13,7 @@ public interface CoinsDropService {
 
     List<CoinsDropEntity> getByOrganizationId(Long orgId);
 
-    CoinsDropEntity getByOrganizationIdAndTypeId(Long orgId, Long typeId);
+    CoinsDropEntity getByOrganizationIdAndTypeId(Long orgId, Integer typeId);
 
     void giveUsersCoinsBirthDay();
